+++ Beam/resources/assets/js/remplib.js
@@ -448,6 +448,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "funnel_id": funnelId
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.generateRempCheckoutID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -466,6 +467,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.generateRempCheckoutID(),
             };
 
             params = this.addSystemUserParams(params);
@@ -487,6 +489,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "product_ids": productIds
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.getRempCheckoutID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -509,6 +512,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.getRempCheckoutID(),
             };
 
             params = this.addSystemUserParams(params);
@@ -530,6 +534,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "product_ids": productIds
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.getRempCheckoutID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -552,6 +557,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "source": source
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.getRempCheckoutID(),
             };
 
             params = this.addSystemUserParams(params);
@@ -573,6 +579,7 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                     "product_ids": productIds
                 },
                 "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.getRempCheckoutID(),
             };
             params = this.addSystemUserParams(params);
             this.post(this.url + "/track/commerce", params);
@@ -594,7 +601,8 @@ remplib = typeof(remplib) === 'undefined' ? {} : remplib;
                 "user": {
                     "source": source
                 },
-                "remp_commerce_id": remplib.uuidv4()
+                "remp_commerce_id": remplib.uuidv4(),
+                "remp_checkout_id": remplib.getRempCheckoutID(),
             };
             params = this.addSystemUserParams(params);
             params["user"]["source"] = source;